+++ app/actions/addresses-actions.js
@@ -4,5 +4,5 @@ import Action from './lib/Action';
 // ======= ADDRESSES ACTIONS =======
 
 export default class AddressesActions {
-  createNew: Action<{ data: Object }> = new Action();
+  createAddress: Action<{ walletId: string, password: ?string }> = new Action();
 }