+++ app/containers/wallet/WalletReceivePage.js
@@ -24,8 +24,23 @@ export default class WalletReceivePage extends Component {
   static defaultProps = { actions: null, stores: null };
   props: InjectedProps;
 
+  state = {
+    copiedAddress: '',
+  };
+
   componentWillUnmount() {
     this.closeNotification();
+  }
+
+  handleGenerateAddress = () => {
+    const { wallets } = this.props.stores;
+    const wallet = wallets.active;
+    if (wallet) {
+      this.props.actions.addresses.createAddress.trigger({
+        walletId: wallet.id,
+        password: null, // TODO: there is no password input in the UI
+      });
+    }
   }
 
   closeNotification = () => {
@@ -38,14 +53,16 @@ export default class WalletReceivePage extends Component {
   };
 
   render() {
+    const { copiedAddress } = this.state;
     const actions = this.props.actions;
-    const { wallets, addresses, uiNotifications } = this.props.stores;
+    const { wallets, addresses, sidebar, uiNotifications } = this.props.stores;
     const wallet = wallets.active;
 
     // Guard against potential null values
     if (!wallet) throw new Error('Active wallet required for WalletReceivePage.');
 
     const walletAddress = addresses.active ? addresses.active.id : '';
+    const walletAddresses = addresses.all.reverse();
 
     const notification = {
       id: `${wallet.id}-copyNotification`,
@@ -53,7 +70,7 @@ export default class WalletReceivePage extends Component {
       message: (
         <FormattedHTMLMessage
           {...messages.message}
-          values={{ walletAddress: ellipsis(walletAddress, 8) }}
+          values={{ walletAddress: ellipsis(copiedAddress, 8) }}
         />
       ),
     };
@@ -62,14 +79,17 @@ export default class WalletReceivePage extends Component {
       <VerticalFlexContainer>
 
         <WalletReceive
-          walletName={wallet.name}
           walletAddress={walletAddress}
-          onCopyAddress={() => {
+          walletAddresses={walletAddresses}
+          onGenerateAddress={this.handleGenerateAddress}
+          onCopyAddress={(address) => {
+            this.setState({ copiedAddress: address });
             actions.notifications.open.trigger({
               id: notification.id,
               duration: notification.duration,
             });
           }}
+          isSidebarExpanded={sidebar.isShowingSubMenus}
         />
 
         <NotificationMessage