+++ app/stores/AddressesStore.js
@@ -3,8 +3,12 @@ import { observable, computed, action } from 'mobx';
 import _ from 'lodash';
 import Store from './lib/Store';
 import CachedRequest from './lib/LocalizedCachedRequest';
+import Request from './lib/LocalizedRequest';
 import WalletAddress from '../domain/WalletAddress';
-import type { GetAddressesResponse } from '../api';
+import type {
+  GetAddressesResponse,
+  CreateAddressResponse,
+} from '../api';
 
 export default class AddressesStore extends Store {
 
@@ -13,9 +17,28 @@ export default class AddressesStore extends Store {
     allRequest: CachedRequest<GetAddressesResponse>
   }> = [];
 
+  // REQUESTS
+  /* eslint-disable max-len */
+  @observable createAddressRequest: Request<CreateAddressResponse> = new Request(this.api.createAddress);
+  /* eslint-disable max-len */
+
   setup() {
-    // const actions = this.actions.addresses;
+    const actions = this.actions.addresses;
+    actions.createAddress.listen(this._createAddress);
   }
+
+  _createAddress = async (params: { walletId: string, password: ?string }) => {
+    try {
+      const { walletId, password } = params;
+      const accountId = this._getAccountIdByWalletId(walletId);
+      const address: ?CreateAddressResponse = await this.createAddressRequest.execute({
+        accountId, password
+      }).promise;
+      if (address != null) this._refreshAddresses();
+    } catch (error) {
+      throw error;
+    }
+  };
 
   @computed get all(): Array<WalletAddress> {
     const wallet = this.stores.wallets.active;